refactor: update generateRe... params
Having too many params and strict requirements to perform in a way it is supposed to do makes harder to use it.

I removed all unncessary code and expect the all code is in the refreshTokenRequestFunc.

Calling the refreshTokenRequestFunc is only thing the link does, it is simpler than before as I think.

in the old code, refreshToken catches an error in the middle and it passes a request that failed to get a refresh token.

I removed the `try-block` code and if the refreshToken occurs errors then the link will be broken, which means that failed operations don't retry.